Move OsLoginAdapter to Apis project
