+++ sources/Google.Solutions.Apis.Test/packages.config
@@ -3,6 +3,7 @@
   <package id="Castle.Core" version="5.1.1" targetFramework="net462" />
   <package id="Google.Apis" version="1.60.0" targetFramework="net462" />
   <package id="Google.Apis.Auth" version="1.60.0" targetFramework="net462" />
+  <package id="Google.Apis.CloudOSLogin.v1" version="1.60.0.2634" targetFramework="net462" />
   <package id="Google.Apis.CloudResourceManager.v1" version="1.60.0.2950" targetFramework="net462" />
   <package id="Google.Apis.Compute.v1" version="1.60.0.3050" targetFramework="net462" />
   <package id="Google.Apis.Core" version="1.60.0" targetFramework="net462" />
