+++ sources/Google.Solutions.Apis/Compute/OsLoginAdapter.cs
@@ -21,23 +21,18 @@
 
 using Google.Apis.CloudOSLogin.v1;
 using Google.Apis.CloudOSLogin.v1.Data;
-using Google.Solutions.Apis;
 using Google.Solutions.Apis.Auth;
 using Google.Solutions.Apis.Client;
 using Google.Solutions.Apis.Locator;
 using Google.Solutions.Common.Diagnostics;
 using Google.Solutions.Common.Util;
-using Google.Solutions.IapDesktop.Application;
-using Google.Solutions.IapDesktop.Core.ObjectModel;
-using Google.Solutions.IapDesktop.Application.Services.Adapters;
-using Google.Solutions.Ssh.Auth;
 using System;
 using System.Diagnostics;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Google.Solutions.IapDesktop.Extensions.Session.Protocol.Adapter
+namespace Google.Solutions.Apis.Compute
 {
     public interface IOsLoginAdapter
     {
@@ -68,7 +63,6 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Protocol.Adapter
             CancellationToken cancellationToken);
     }
 
-    [Service(typeof(IOsLoginAdapter))]
     public class OsLoginAdapter : IOsLoginAdapter
     {
         private const string MtlsBaseUri = "https://oslogin.mtls.googleapis.com/";
@@ -111,7 +105,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Protocol.Adapter
 
             Debug.Assert(!keyType.Contains(' '));
 
-            using (ApplicationTraceSources.Default.TraceMethod().WithParameters(project))
+            using (ApiTraceSources.Default.TraceMethod().WithParameters(project))
             {
                 var expiryTimeUsec = new DateTimeOffset(DateTime.UtcNow.Add(validity))
                     .ToUnixTimeMilliseconds() * 1000;
@@ -181,7 +175,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Protocol.Adapter
             ProjectLocator project,
             CancellationToken token)
         {
-            using (ApplicationTraceSources.Default.TraceMethod().WithParameters(project))
+            using (ApiTraceSources.Default.TraceMethod().WithParameters(project))
             {
                 var request = this.service.Users.GetLoginProfile(
                     $"users/{this.authorization.Email}");
@@ -208,7 +202,7 @@ namespace Google.Solutions.IapDesktop.Extensions.Session.Protocol.Adapter
             string fingerprint,
             CancellationToken cancellationToken)
         {
-            using (ApplicationTraceSources.Default.TraceMethod().WithParameters(fingerprint))
+            using (ApiTraceSources.Default.TraceMethod().WithParameters(fingerprint))
             {
                 try
                 {
