+++ sources/Google.Solutions.Apis/Google.Solutions.Apis.csproj
@@ -41,6 +41,9 @@
     <Reference Include="Google.Apis.Auth.PlatformServices, Version=1.60.0.0, Culture=neutral, PublicKeyToken=4b01fa6e34db77ab, processorArchitecture=MSIL">
       <HintPath>..\packages\Google.Apis.Auth.1.60.0\lib\net461\Google.Apis.Auth.PlatformServices.dll</HintPath>
     </Reference>
+    <Reference Include="Google.Apis.CloudOSLogin.v1, Version=1.60.0.2634, Culture=neutral, PublicKeyToken=4b01fa6e34db77ab, processorArchitecture=MSIL">
+      <HintPath>..\packages\Google.Apis.CloudOSLogin.v1.1.60.0.2634\lib\net45\Google.Apis.CloudOSLogin.v1.dll</HintPath>
+    </Reference>
     <Reference Include="Google.Apis.CloudResourceManager.v1, Version=1.60.0.2950, Culture=neutral, PublicKeyToken=4b01fa6e34db77ab, processorArchitecture=MSIL">
       <HintPath>..\packages\Google.Apis.CloudResourceManager.v1.1.60.0.2950\lib\net45\Google.Apis.CloudResourceManager.v1.dll</HintPath>
     </Reference>
@@ -74,6 +77,7 @@
     <Compile Include="Client\RestClient.cs" />
     <Compile Include="Client\UserAgent.cs" />
     <Compile Include="Compute\ComputeEngineAdapter.cs" />
+    <Compile Include="Compute\OsLoginAdapter.cs" />
     <Compile Include="HelpTopics.cs" />
     <Compile Include="Compute\IComputeEngineAdapter.cs" />
     <Compile Include="Compute\AddressResolver.cs" />
