+++ sources/Google.Solutions.IapDesktop.Extensions.Session.Test/Protocol/Ssh/TestOsLoginService.cs
@@ -20,6 +20,7 @@
 //
 
 using Google.Apis.CloudOSLogin.v1.Data;
+using Google.Solutions.Apis.Compute;
 using Google.Solutions.Apis.Locator;
 using Google.Solutions.Common.Util;
 using Google.Solutions.IapDesktop.Extensions.Session.Protocol.Adapter;
