+++ sources/Google.Solutions.IapDesktop/Program.cs
@@ -462,6 +462,7 @@ namespace Google.Solutions.IapDesktop
                 mainLayer.AddSingleton<IResourceManagerAdapter, ResourceManagerAdapter>();
                 mainLayer.AddSingleton<IComputeEngineAdapter, ComputeEngineAdapter>();
                 mainLayer.AddTransient<IAddressResolver, AddressResolver>();
+                mainLayer.AddSingleton<IOsLoginAdapter, OsLoginAdapter>();
 
                 mainLayer.AddTransient<IWindowsCredentialGenerator, WindowsCredentialGenerator>();
                 mainLayer.AddSingleton<IJobService, JobService>();
